Added: company table

+++ Models/Employee.cs
@@ -1,8 +1,10 @@
+using System.Runtime.InteropServices.JavaScript;
+
 namespace MinimalApi.Models;
 
 public class Employee
 {
-    public int Id { get; set; }
+    public Int64 Id { get; set; }
     public string? FirstName { get; set; }
     public string? LastName { get; set; }
     public string? Branch { get; set; }
